+++ test/01-migrator/configure/adapterLoading.js
@@ -16,35 +16,36 @@ describe('migrator configure adapter loading', () => {
 		Adapter.prototype = testUtils.createAdapter({withCallbacMethods: true});
 	});
 
-	it('should try migrator-related path first then CWD-related', () => {
+	it('should load module related to cwd', () => {
 		const paths = [];
 
 		return Promise.resolve()
 			.then(() => {
 				const migrator = new Migrator();
 
-				migrator._tryLoadModule = (path) => {
+				migrator._loadModule = (path) => {
 					paths.push(path);
 
-					return paths.length === 2 ?
-						Promise.resolve(Adapter) :
-						Promise.reject(new Error('Whatever.'));
+					if (paths.length === 1) {
+						return Promise.resolve(Adapter);
+					} else {
+						return Promise.reject(new Error('Whatever.'));
+					}
 				};
 
 				return migrator.configure({adapter: 'X', loadConfig: false});
 			})
 			.then(() => {
-				expect(paths[0]).eql('X');
-				expect(paths[1].substr(-2, 2)).eql('/X');
+				expect(paths[0].substr(-2, 2)).eql('/X');
 			});
 	});
 
-	it('should throw an error when both paths cannot be resolved', () => {
+	it('should throw an error when cannot load module', () => {
 		return Promise.resolve()
 			.then(() => {
 				const migrator = new Migrator();
 
-				migrator._tryLoadModule = () => Promise.reject(new Error('Whatever.'));
+				migrator._loadModule = () => Promise.reject(new Error('Whatever.'));
 
 				return migrator.configure({adapter: 'X', loadConfig: false});
 			})
@@ -55,14 +56,14 @@ describe('migrator configure adapter loading', () => {
 				expect(err).ok();
 				expect(err).an(Error);
 				expect(err.message).match(
-					/Error loading adapter from all paths:\s+Error: Whatever./
+					/Error during adapter creation: Whatever\./
 				);
 			});
 	});
 
 	it('should promisify adapter methods', () => {
 		const migrator = new Migrator();
-		migrator._tryLoadModule = () => Promise.resolve(Adapter);
+		migrator._loadModule = () => Promise.resolve(Adapter);
 		const adapterMethodNames = [
 			'connect',
 			'disconnect',
@@ -117,7 +118,7 @@ describe('migrator configure adapter loading', () => {
 				expect(err).ok();
 				expect(err).an(Error);
 				expect(err.message).equal(
-					'Error constructing adapter:Some error'
+					'Error during adapter creation: Some error'
 				);
 			});
 	});