+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and `MigrationManager` class)
 - Builtin TypeScript migration template file
 
 ### Changed
+- *Breaking change:* Adapter/plugin path resolves related to cwd only (earlier
+tried to resolve migrator related path first then cwd related)
 - *Breaking change:* Default migration template now uses `async` functions
 - *Breaking change:* `MigrationManager` `onSkipMigration` event reasons are
 renamed: canNotMigrateAlreadyExecuted -> cannotMigrateAlreadyExecuted,